+++ src/main/java/org/ramer/diary/domain/Topic.java
@@ -1,15 +1,15 @@
 package org.ramer.diary.domain;
 
-import java.io.Serializable;
-import java.util.Date;
-import java.util.HashSet;
-import java.util.Set;
-
-import javax.persistence.*;
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import java.io.Serializable;
+import java.util.Date;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Set;
 
 /**
  * 分享类.
@@ -43,9 +43,9 @@ public class Topic implements Serializable{
     private String picture;
 
     //  标签
-    @ManyToOne
-    @JoinColumn(name = "tags")
-    private Tags tags;
+    @OrderBy("hot desc")
+    @OneToMany(cascade = { CascadeType.REMOVE }, mappedBy = "topic")
+    private List<Tags> tagses;
 
     /** 点赞次数. */
     @Column
@@ -74,7 +74,7 @@ public class Topic implements Serializable{
     @Override
     public String toString() {
         return "Topic{" + "id=" + id + ", content='" + content + '\'' + ", date=" + date + ", picture='" + picture
-                + '\'' + ", tags='" + tags + '\'' + ", upCounts=" + upCounts + '}';
+                + '\'' + ", upCounts=" + upCounts + '}';
     }
 
     @Override
@@ -96,8 +96,6 @@ public class Topic implements Serializable{
             return false;
         if (picture != null ? !picture.equals(topic.picture) : topic.picture != null)
             return false;
-        if (tags != null ? !tags.equals(topic.tags) : topic.tags != null)
-            return false;
         return upCounts != null ? upCounts.equals(topic.upCounts) : topic.upCounts == null;
     }
 
@@ -108,7 +106,6 @@ public class Topic implements Serializable{
         result = 31 * result + (content != null ? content.hashCode() : 0);
         result = 31 * result + (date != null ? date.hashCode() : 0);
         result = 31 * result + (picture != null ? picture.hashCode() : 0);
-        result = 31 * result + (tags != null ? tags.hashCode() : 0);
         result = 31 * result + (upCounts != null ? upCounts.hashCode() : 0);
         return result;
     }
