+++ src/main/java/org/ramer/diary/service/impl/TagsServiceImpl.java
@@ -1,9 +1,19 @@
 package org.ramer.diary.service.impl;
 
+import org.ramer.diary.domain.Tags;
 import org.ramer.diary.service.TagsService;
 
 /**
  * Created by caipeijun on 2017/5/29.
  */
 public class TagsServiceImpl implements TagsService {
+    @Override
+    public Tags findOne(Integer id) {
+        return null;
+    }
+
+    @Override
+    public Tags findOneByName(String name) {
+        return null;
+    }
 }
