+++ src/main/resources/static/css/component.css
@@ -9,7 +9,7 @@
  display: block;
  float: left;
  padding: 10px;
- width: 33%;
+ width: 33.3%;
  margin-bottom: 15px;
 }
