+++ src/main/resources/templates/home.html
@@ -213,7 +213,7 @@
                 <img alt="" src="/pictures/calendar.png"/>
             </div>
         </div>
-        <input class="upbtn" type="file" name="picture" accept="image/*" id="upPic"/>
+        <input class="upbtn" type="file" name="picture" accept="image/*" id="upPic" multiple="multiple"/>
         <input class="upbtn2" placeholder="点击添加一张图片"/>
         <div id="preview" class="preview"></div>
         <div class="input_tag_panel">
@@ -225,6 +225,14 @@
         <input type="submit" value="分享" id="submitTopic"/>
         <input type="reset" value="收起" class="hiddenTopic"/>
     </form>
+</div>
+<!-- 上一页 -->
+<div class="last_page">
+    <th:block th:if="${session.scrollInPage eq false}">
+        <a th:href="|?pageNum=${topics.number + 1 - 1}|" id="lastPage">
+            <img alt="error" src="/pictures/previous.png"/>
+        </a>
+    </th:block>
 </div>
 <!-- 显示分享 -->
 <div class="container" th:if="${showTopic eq 'true' }">
@@ -259,26 +267,19 @@
     <!-- 存储当前页号 -->
     <input id="number" type="hidden" th:value="${topics.number}"/>
 </div>
-<!-- 页码面板 -->
-<ul class="page_panel">
+<!-- 下一页 -->
+<div class="next_page">
     <th:block th:if="${session.scrollInPage eq false}">
-        <li class="lastPage">
-            <a th:href="|?pageNum=${topics.number + 1 - 1}|" id="lastPage">
-                <img alt="error" src="/pictures/previous.png"/>
-            </a>
-        </li>
-        <li class="nextPage">
-            <a th:href="|?pageNum=${topics.number + 1 + 1}|" id="nextPage">
-                <img alt="error" src="/pictures/next.png"/>
-            </a>
-        </li>
-    </th:block>
-    <li>
-        <a id="toggleScrollInPage" th:href="'/scrollInPage?'+${_csrf.parameterName}+'=' +${_csrf.token}">
-            <img alt="error" src="/pictures/option.png"/>
+        <a th:href="|?pageNum=${topics.number + 1 + 1}|" id="nextPage">
+            <img alt="error" src="/pictures/next.png"/>
         </a>
-    </li>
-</ul>
+    </th:block>
+    <!--<li>-->
+    <!--<a id="toggleScrollInPage" th:href="'/scrollInPage?'+${_csrf.parameterName}+'=' +${_csrf.token}">-->
+    <!--<img alt="error" src="/pictures/option.png"/>-->
+    <!--</a>-->
+    <!--</li>-->
+</div>
 <!-- 显示达人 -->
 <div class="container" th:if="${showTopPeople eq 'true' }">
     <ul class="grid effect-2" id="grid">
